# [Non-breaking changes] Chore: Update derive SHA for stacktrie