--- a/consensus/clique/clique.go
+++ b/consensus/clique/clique.go
@@ -576,7 +576,7 @@ func (c *Clique) Finalize(chain consensus.ChainReader, header *types.Header, sta
 	header.UncleHash = types.CalcUncleHash(nil)
 
 	// Assemble and return the final block for sealing
-	return types.NewBlock(header, txs, nil, receipts, new(trie.Trie)), nil
+	return types.NewBlock(header, txs, nil, receipts, new(trie.StackTrie)), nil
 }
 
 // Authorize injects a private key into the consensus engine to mint new blocks