--- a/consensus/ethash/consensus.go
+++ b/consensus/ethash/consensus.go
@@ -520,7 +520,7 @@ func (ethash *Ethash) Finalize(chain consensus.ChainReader, header *types.Header
 	header.Root = state.IntermediateRoot(chain.Config().IsEIP158(header.Number))
 
 	// Header seems complete, assemble into a block and return
-	return types.NewBlock(header, txs, uncles, receipts, new(trie.Trie)), nil
+	return types.NewBlock(header, txs, uncles, receipts, new(trie.StackTrie)), nil
 }
 
 // Some weird constants to avoid constant memory allocs for them.