--- a/consensus/posv/posv.go
+++ b/consensus/posv/posv.go
@@ -986,7 +986,7 @@ func (c *Posv) Finalize(chain consensus.ChainReader, header *types.Header, state
 	header.UncleHash = types.CalcUncleHash(nil)
 
 	// Assemble and return the final block for sealing
-	return types.NewBlock(header, txs, nil, receipts, new(trie.Trie)), nil
+	return types.NewBlock(header, txs, nil, receipts, new(trie.StackTrie)), nil
 }
 
 // Authorize injects a private key into the consensus engine to mint new blocks