--- a/core/genesis.go
+++ b/core/genesis.go
@@ -260,7 +260,7 @@ func (g *Genesis) ToBlock(db ethdb.Database) *types.Block {
 	statedb.Commit(false)
 	statedb.Database().TrieDB().Commit(root, true)
 
-	return types.NewBlock(head, nil, nil, nil, new(trie.Trie))
+	return types.NewBlock(head, nil, nil, nil, new(trie.StackTrie))
 }
 
 // Commit writes the block and state of a genesis specification to the database.