--- a/core/types/hashing.go
+++ b/core/types/hashing.go
@@ -1,0 +1,11 @@
+package types
+
+import (
+	"bytes"
+	"sync"
+)
+
+// encodeBufferPool holds temporary encoder buffers for DeriveSha and TX encoding.
+var encodeBufferPool = sync.Pool{
+	New: func() interface{} { return new(bytes.Buffer) },
+}