--- a/core/types/receipt.go
+++ b/core/types/receipt.go
@@ -206,11 +206,9 @@ type Receipts []*Receipt
 // Len returns the number of receipts in this list.
 func (r Receipts) Len() int { return len(r) }
 
-// GetRlp returns the RLP encoding of one receipt from the list.
-func (r Receipts) GetRlp(i int) []byte {
-	bytes, err := rlp.EncodeToBytes(r[i])
-	if err != nil {
-		panic(err)
-	}
-	return bytes
+// EncodeIndex encodes the i'th receipt to w.
+func (rs Receipts) EncodeIndex(i int, w *bytes.Buffer) {
+	r := rs[i]
+	data := &receiptRLP{r.statusEncoding(), r.CumulativeGasUsed, r.Bloom, r.Logs}
+	rlp.Encode(w, data)
 }