This PR update deriveSha function for stacktrie. DeriveSha creates the tree hashes of transactions, receipts, and withdrawals in a block header.
- Need merge this PR first: https://github.com/tomochain/tomochain/pull/370
- Benchmark after use stacktrie alternative for trie: 
``` 
goos: darwin
goarch: amd64
pkg: github.com/tomochain/tomochain/core/types
cpu: VirtualApple @ 2.50GHz
                        │     old     │                new                 │
                        │   sec/op    │   sec/op     vs base               │
DeriveSha200/std_trie-8   215.6µ ± 2%   198.2µ ± 1%  -8.04% (p=0.000 n=20)

                        │      old      │                 new                  │
                        │     B/op      │     B/op      vs base                │
DeriveSha200/std_trie-8   279.85Ki ± 0%   61.23Ki ± 0%  -78.12% (p=0.000 n=20)

                        │     old     │                 new                 │
                        │  allocs/op  │  allocs/op   vs base                │
DeriveSha200/std_trie-8   2.888k ± 0%   1.647k ± 0%  -42.97% (p=0.000 n=20)
```