@trinhdn2 2023-07-25
This PR can be marked as non-breaking, and the merge destination needs to be changed from `tomochain:master` to `tomochain:upgrade-core-develop`